Update forc-crypto to use latest libraries
