+++ forc-plugins/forc-client/src/util/tx.rs
@@ -313,4 +313,3 @@ impl<T: field::Witnesses> TransactionExt for T {
         self
     }
 }
-
