--- forc-plugins/forc-crypto/src/keygen/mod.rs
@@ -1,20 +1,0 @@
-use clap::ValueEnum;
-
-pub mod new_key;
-pub mod parse_secret;
-
-#[derive(Clone, Debug, Default, ValueEnum)]
-pub enum KeyType {
-    #[default]
-    BlockProduction,
-    Peering,
-}
-
-impl From<KeyType> for &'static str {
-    fn from(key_type: KeyType) -> Self {
-        match key_type {
-            KeyType::BlockProduction => "block-production",
-            KeyType::Peering => "p2p",
-        }
-    }
-}
